Remove lateral scrolling scaffolding.
I wasn't sure if the niri windows IPC update was going to be sufficient
for my purposes here - tl;dr it isn't (no events emitted when a column
is centered -> not enough information to deduce when to scroll). So, I'm
removing all that jazz.

Because I had already implemented the relevant bits on the render side
(and had previously stubbed out the Niri side to just default-value 50%
for the scroll every time on the x-axis), I used claude to munge all the
window IPC bits for me to evaluate if it would be a sufficiently
pleasing experience. It wasn't, so I cleaned up all the resulting work.